Add a stub file for a refactoring exercise
This stub file's main section contains function signatures that
describe how street_crossing.py could be refactored to use functions
instead of being a straight script.

The exercise is to "fill in the blank" (designated by the ###) so that
this stub has the same output as street_crossing.py.

Changes can be made in this file, or street_crossing.py.